1、网络日志，划分请求体和响应体打印 2、网络日志，本地化存储，新增url和method字段记录

+++ library_log/src/main/java/com/venson/versatile/log/database/entity/LogEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class LogEntity(
 
     //自增主键，构造函数赋值为null
-    @PrimaryKey(autoGenerate = true) var id: Int?,
+    @PrimaryKey(autoGenerate = true) var id: Long?,
 
     var time: Long,
